echo_srv: Check for a valid pidfile pointer

+++ tests/echo_srv.c
@@ -916,7 +916,7 @@ int main(int argc, char **argv)
     echo(sock, &opts);
     close(sock);
 
-    if (opts.daemon) {
+    if (opts.daemon && opts.pidfile != NULL) {
         unlink(opts.pidfile);
     }
